NBAProject/requirements.txt
scrapy
scrapy-djangoitem
beautifulsoup4
django
djangorestframework
netifaces